requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-autoslug==1.9.9
django-extensions==3.2.3
django-rosetta==0.9.9
django-widget-tweaks==1.5.0
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-socketio==0.0.10
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
markdown2==2.4.10
MarkupSafe==2.1.3
multidict==6.0.4
nicegui==1.3.15
nltk==3.8.1
orjson==3.9.7
Pillow==10.0.1
polib==1.2.0
pscript==0.7.7
psycopg2==2.9.7
pydantic==1.10.13
Pygments==2.16.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-multipart==0.0.6
python-socketio==5.9.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.5
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
watchfiles==0.20.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
